feat(Footer): extract footer rendering logic into FooterClient component for better separation of concerns feat(FullBleed): optimize image loading with higher resolution and quality settings feat(LandingModule): enhance image loading with priority and improved quality settings
